requirements.txt

paho-mqtt==1.6.1
pymongo==3.12.0
fastapi==0.68.1
uvicorn==0.15.0
redis==3.5.3
pydantic==1.8.3

